Adapting folder structure for models and embeddings
This restructuring is made in preparation for the implementation of the ability to choose between LangChain or LlamaIndex as main frameworks.
