+++ src/embeddings/langChain/ada-embeddings.ts
@@ -1,5 +1,5 @@
 import { OpenAIEmbeddings } from '@langchain/openai';
-import { BaseEmbeddings } from '../interfaces/base-embeddings.js';
+import { BaseEmbeddings } from '../../interfaces/base-embeddings.js';
 
 export class AdaEmbeddings implements BaseEmbeddings {
     private model: OpenAIEmbeddings;
