+++ src/embeddings/langChain/fireworks-embeddings.ts
@@ -1,5 +1,5 @@
 import { FireworksEmbeddings } from "@langchain/community/embeddings/fireworks";
-import { BaseEmbeddings } from '../interfaces/base-embeddings.js';
+import { BaseEmbeddings } from '../../interfaces/base-embeddings.js';
 
 export class FireworksEmbedding implements BaseEmbeddings {
     private model: FireworksEmbeddings;
