+++ src/embeddings/langChain/nomic-v1-5-embeddings.ts
@@ -1,5 +1,5 @@
 import { FireworksEmbeddings } from "@langchain/community/embeddings/fireworks";
-import { BaseEmbeddings } from '../interfaces/base-embeddings.js';
+import { BaseEmbeddings } from '../../interfaces/base-embeddings.js';
 
 export class NomicEmbeddingsv1_5 implements BaseEmbeddings {
     private model: FireworksEmbeddings;
