+++ src/models/langChain/bedrock-model.ts
@@ -1,46 +1,33 @@
 import createDebugMessages from 'debug';
-// import { AzureOpenAI } from "@langchain/openai";
-// import { AzureChatOpenAI } from "@langchain/azure-openai";
-import { AzureChatOpenAI } from "@langchain/openai";
-// import { ChatOpenAI } from "@langchain/openai";
-
+import { BedrockChat } from "@langchain/community/chat_models/bedrock";
 import { HumanMessage, AIMessage, SystemMessage } from '@langchain/core/messages';
 
-import { BaseModel } from '../interfaces/base-model.js';
-import { Chunk, ConversationHistory } from '../global/types.js';
+import { BaseModel } from '../../interfaces/base-model.js';
+import { Chunk, ConversationHistory } from '../../global/types.js';
 import { StringOutputParser } from '@langchain/core/output_parsers';
 
-export class AzureChatAI extends BaseModel {
+export class Bedrock extends BaseModel {
 
-    private readonly debug = createDebugMessages('maap:model:AzureOpenAI');
-    private readonly azureOpenAIApiInstanceName: string;
-    private readonly azureOpenAIApiDeploymentName: string;
-    private model: AzureChatOpenAI;
-    private readonly azureOpenAIApiVersion: string;
+    private readonly debug = createDebugMessages('maap:model:Bedrock');
+    private readonly modelName: string;
     private readonly maxTokens: number;
-    private readonly topP: number;
-    private readonly topK: number;
+    private model: BedrockChat;
 
-    constructor(params?: { azureOpenAIApiInstanceName?: string; azureOpenAIApiDeploymentName?: string; azureOpenAIApiVersion?: string; modelName?: string; temperature?: number; maxTokens?: number; topP?:number; topK?:number }) {
+    constructor(params?: { modelName?: string;  maxTokens?: number; temperature?: number }) {
         super(params?.temperature ?? 0.1);
-        this.azureOpenAIApiInstanceName = params?.azureOpenAIApiInstanceName;
-        this.azureOpenAIApiDeploymentName = params?.azureOpenAIApiDeploymentName;
-        this.azureOpenAIApiVersion = params?.azureOpenAIApiVersion;
+        this.modelName = params?.modelName;
         this.maxTokens = params?.maxTokens ?? 2048;
-        this.topP = params?.topP ?? 0.9;
-        this.topK = params?.topK ?? 40;
-
     }
 
     override async init(): Promise<void> {
-        this.model = new AzureChatOpenAI({
-            azureOpenAIApiInstanceName: this.azureOpenAIApiInstanceName,
-            azureOpenAIApiDeploymentName: this.azureOpenAIApiDeploymentName,
-            azureOpenAIApiVersion: this.azureOpenAIApiVersion,
-            temperature: this.temperature,
+        this.model = new BedrockChat({ model: this.modelName, 
             maxTokens: this.maxTokens,
-            topP: this.topP
-          });
+            region: process.env.BEDROCK_AWS_REGION!,
+            credentials: {
+              accessKeyId: process.env.BEDROCK_AWS_ACCESS_KEY_ID!,
+              secretAccessKey: process.env.BEDROCK_AWS_SECRET_ACCESS_KEY!,
+            }
+        });
     }
 
     override async runQuery(
@@ -51,7 +38,7 @@ export class AzureChatAI extends BaseModel {
     ): Promise<string> {
         const pastMessages: (AIMessage | SystemMessage | HumanMessage)[] = this.generatePastMessages(system, supportingContext, pastConversations, userQuery);
         const result = await this.model.invoke(pastMessages);
-        this.debug('AzureOpenAI response -', result);
+        this.debug('Bedrock response -', result);
         return result.content.toString();
     }
 
@@ -78,7 +65,7 @@ export class AzureChatAI extends BaseModel {
         );
         pastMessages.push(new HumanMessage(`${userQuery}?`));
 
-        this.debug('Executing AzureOpenAI model with prompt -', userQuery);
+        this.debug('Executing Bedrock model with prompt -', userQuery);
         return pastMessages;
     }
