+++ src/models/langChain/cohere-model.ts
@@ -3,8 +3,8 @@ import { ChatCohere } from '@langchain/cohere';
 import { HumanMessage, AIMessage, SystemMessage } from '@langchain/core/messages';
 import { StringOutputParser } from '@langchain/core/output_parsers';
 
-import { BaseModel } from '../interfaces/base-model.js';
-import { Chunk, ConversationHistory } from '../global/types.js';
+import { BaseModel } from '../../interfaces/base-model.js';
+import { Chunk, ConversationHistory } from '../../global/types.js';
 
 export class Cohere extends BaseModel {
     private readonly debug = createDebugMessages('maap:model:Cohere');
