+++ src/models/langChain/ollama-model.ts
@@ -1,25 +1,23 @@
 import createDebugMessages from 'debug';
-import { ChatOpenAI } from '@langchain/openai';
-import { HumanMessage, AIMessage, SystemMessage } from '@langchain/core/messages';
+import { Ollama as ChatOllamaAI } from '@langchain/community/llms/ollama';
+import { AIMessage, HumanMessage, SystemMessage } from '@langchain/core/messages';
 
-import { BaseModel } from '../interfaces/base-model.js';
-import { Chunk, ConversationHistory } from '../global/types.js';
+import { Chunk, ConversationHistory } from '../../global/types.js';
+import { BaseModel } from '../../interfaces/base-model.js';
 
-export class OpenAi extends BaseModel {
+export class Ollama extends BaseModel {
     protected runStreamQuery(system: string, userQuery: string, supportingContext: Chunk[], pastConversations: ConversationHistory[]): Promise<any> {
         throw new Error('Method not implemented.');
     }
-    private readonly debug = createDebugMessages('maap:model:OpenAi');
-    private readonly modelName: string;
-    private model: ChatOpenAI;
+    private readonly debug = createDebugMessages('maap:model:Ollama');
+    private model: ChatOllamaAI;
 
-    constructor(params?: { temperature?: number; modelName?: string; }) {
-        super(params?.temperature ?? 0.1);
-        this.modelName = params?.modelName ?? 'gpt-3.5-turbo';
-    }
-
-    override async init(): Promise<void> {
-        this.model = new ChatOpenAI({ temperature: this.temperature, model: this.modelName });
+    constructor({ baseUrl, temperature, modelName }: { baseUrl?: string; temperature?: number; modelName?: string }) {
+        super(temperature);
+        this.model = new ChatOllamaAI({
+            model: modelName ?? 'llama2',
+            baseUrl: baseUrl ?? 'http://localhost:11434',
+        });
     }
 
     override async runQuery(
@@ -43,9 +41,9 @@ export class OpenAi extends BaseModel {
         );
         pastMessages.push(new HumanMessage(`${userQuery}?`));
 
-        this.debug('Executing openai model with prompt -', userQuery);
+        this.debug(`Executing ollama model ${this.model} with prompt -`, userQuery);
         const result = await this.model.invoke(pastMessages);
-        this.debug('OpenAI response -', result);
-        return result.content.toString();
+        this.debug('Ollama response -', result);
+        return result.toString();
     }
 }
