+++ src/models/langChain/togetherai-model.ts
@@ -3,8 +3,8 @@ import { ChatTogetherAI } from '@langchain/community/chat_models/togetherai';
 import { HumanMessage, AIMessage, SystemMessage } from '@langchain/core/messages';
 import { StringOutputParser } from '@langchain/core/output_parsers';
 
-import { BaseModel } from '../interfaces/base-model.js';
-import { Chunk, ConversationHistory } from '../global/types.js';
+import { BaseModel } from '../../interfaces/base-model.js';
+import { Chunk, ConversationHistory } from '../../global/types.js';
 
 export class TogetherAI extends BaseModel {
     private readonly debug = createDebugMessages('maap:model:TogetherAI');
